+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import type { StoryObj, Meta } from '@storybook/react'
 
-import { Box, Text, Checkbox, CheckboxProps } from '@lemonade-technologies-ui/react'
+import { Box, Text, Checkbox, CheckboxProps } from '@lemonade-technologies-hub-ui/react'
 
 export default {
   title: 'React/Form/Checkbox',