+++ packages/docs/src/stories/Text.stories.tsx
@@ -1,6 +1,6 @@
 import type { StoryObj, Meta } from '@storybook/react'
 
-import { Text, TextProps } from '@lemonade-technologies-ui/react'
+import { Text, TextProps } from '@lemonade-technologies-hub-ui/react'
 
 export default {
   title: 'React/Typografy/Text',